it's all done!

+++ src/main/java/org/homework/impl/EmployeeRepository.java
@@ -7,7 +7,7 @@ public interface EmployeeRepository <E>{
 
     void add(E employee);
 
-    E findByExperience (byte expAmount);
+    E findByExperience (int expAmount);
     E findPhoneByName(String phoneNumber);
     E findById(int id);
     ArrayList<E> getAll();
